feature: loadGoogleMaps 함수 분리
