+++ src/utils/loadGoogleMaps.ts
@@ -1,0 +1,22 @@
+import { env } from '../schemas/envSchema.ts';
+
+let googleMapsPromise: Promise<void> | null = null;
+export const loadGoogleMaps = () => {
+  if (googleMapsPromise) return googleMapsPromise;
+
+  googleMapsPromise = new Promise((resolve) => {
+    if (window.google?.maps) {
+      resolve();
+      return;
+    }
+
+    const script = document.createElement('script');
+    script.src = `https://maps.googleapis.com/maps/api/js?key=${env.VITE_GOOGLE_MAPS_API_KEY}&language=ko`;
+    script.async = true;
+    script.defer = true;
+    script.onload = () => resolve();
+    document.head.appendChild(script);
+  });
+
+  return googleMapsPromise;
+};
